requirements.txt
flask
langgraph
langchain
langchain-community
openai
faiss-cpu
sentence-transformers
pdfplumber
opencv-python
pdf2image
python-dotenv
numpy
sentencepiece
thefuzz
python-dateutil
pymupdf
pil